+++ packages/dumb/src/index.ts
@@ -53,6 +53,7 @@ import {
 } from './components/stats/average-daily-dose/average-daily-dose-stat'
 import { SimpleStatMemoized as SimpleStat } from './components/stats/simple/simple-stat'
 import { LabeledCheckbox } from './components/controls/labeled-checkbox/labeled-checkbox'
+import CbgSliceSegment from './components/trends/cbg/cbg-slice-segment'
 
 export {
   AverageDailyDoseStat,
@@ -60,6 +61,7 @@ export {
   BolusTooltip,
   CBGMeanStat,
   CBGPercentageBarChart,
+  CbgSliceSegment,
   CBGStandardDeviation,
   CBGStatType,
   CgmTable,